move into sampler
